Clang format + macOS compilation errors
Inspect and clean up the util library


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error


Fix macOS compilation error

+++ src/genie/core/format-exporter-compressed.cc
@@ -21,7 +21,7 @@ void FormatExporterCompressed::skipIn(const util::Section&) {}
 
 // ---------------------------------------------------------------------------------------------------------------------
 
-void FormatExporterCompressed::flushIn(size_t&) {}
+void FormatExporterCompressed::flushIn(uint64_t&) {}
 
 // ---------------------------------------------------------------------------------------------------------------------
