+++ src/genie/read/spring/encoder-source.h
@@ -56,13 +56,13 @@ class SpringSource : public util::OriginalSource, public util::Source<core::Acce
      * @param lock
      * @return
      */
-    bool pump(size_t& id, std::mutex& lock) override;
+    bool pump(uint64_t& id, std::mutex& lock) override;
 
     /**
      *
      * @param pos
      */
-    void flushIn(size_t& pos) override;
+    void flushIn(uint64_t& pos) override;
 };
 
 // ---------------------------------------------------------------------------------------------------------------------
